Move DATE and VERSION keywords after use statements
The DATE and VERSION keywords get expanded by Dist::Zilla to `our`
variables after building the dist.  With these variables at the top of
the file, this made code appear before strictures were turned on and
hence caused the author critic tests to fail the strict and warnings
checks.  Moving the keywords after the list of dists used by the module
doesn't change the dist's behaviour but allows the author-critic tests
to pass as expected.